+++ R/estimates.R
@@ -72,7 +72,7 @@ get_estimates <- function(geography = c("us", "region", "division", "state", "co
                           shift_geo = FALSE, key = NULL, show_call = FALSE, ...) {
 
   if (missing(vintage) && year > 2020) {
-    rlang::warn(c("`get_estimates()` now uses the `vintage` argument to specify the PEP vintage, and the `year` argument to isolate a year within that vintage.",
+    rlang::warn(c("For post-2020 Census estimates, `get_estimates()` now uses the `vintage` argument to specify the PEP vintage, and the `year` argument to isolate a year within that vintage.",
                   "!" = "This may be a breaking change in your code",
                   "!" = "Omitting `vintage` may lead to incorrect or unexpected results."))
   }