+++ NEWS.md
@@ -1,3 +1,7 @@
+# tidycensus 1.7.0
+
+* Defaults for `get_acs()` and `get_pums()` updated to 2023 following full release of the 2023 ACS datasets.
+
 # tidycensus 1.6.3
 
 * Support added in `get_estimates()` for the 2023 PEP datasets. A new argument, `vintage`, allows users to request data by vintage and then by `year` within that vintage for post-2020 PEP datasets.